Support for multi-target domains
This commit adds ability to have several endpoints for single domain
name with CoreDNS by adding random to the DNS name and setting
TargetStrip to 1 so that it will be stripped upon read. When TXT
record is encountered, it is merged with the first available A
record, if any available. Otherwise dedicated record is created.